Add basic UI with a service worker that caches entry content
The service worker caches content that has been viewed once so that it
can be continue to be read when the server is offline. But, it isn't
very useful for being able to read content that hasn't been read, yet,
which is where the caching might actually be a useful thing.