Simple fetch handler with couple of cache strategies
Use cache_first_with_refresh for shell contents. We use existing shell,
but refresh the cache for the next request.

Use the cache_first strategy for entry content. Content at a particular
hash would never change. New content would have a new hash. So, if we
find a content hash in the cache, we always use it!