Update the UI or notify user when new search results available
Since we return the results from the cache by default, we need a
mechanism to update the UI or inform the user when a backgroun network
request to update search data succeeds. This commit implements that by
submitting a search when the update arrives within 500ms of a cached
result update arriving to the user. In other cases, we show a message to
the user.