Implement offline first updates for tags
Tags are stored in the local state immediately and UI is updated
optimistically. The updates to tags are relayed to the server and they
are attempted to be updated in the background. Also, we update the tags
in the background by simply replaying the updates instead of adding
complex logic to combind all the add/removes.

Closes #4
Closes #27
Closes #12